pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[project]
name = "bsr"
version = "0.1.2"
description = "Blender integration tools for soft-rod system."
readme = "README.md"
authors = []
license = { text = "MIT" }
keywords = ["data-analysis"]
requires-python = ">=3.11,<3.12"
# NOTE: The strict requirement is due to the bpy package
# https://pypi.org/project/bpy/

# Pypi classifiers: https://pypi.org/classifiers/
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Science/Research",
  "Intended Audience :: Education",
  "Operating System :: OS Independent",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
]

dependencies = [
  "bpy==4.5.7",
  "click>=8.1.7,<9",
  "loguru>=0.7.3",
  "mathutils>=3.3.0,<4",
  "pyelastica @ git+https://github.com/GazzolaLab/PyElastica.git@update-v1",
  "websockets>=12,<13",
]

[project.scripts]
# elastica data converter
elastica-npz2blend = "elastica_blender.converter.npz2blend:main"

[dependency-groups]
dev = [
  "black>=24.4.2,<27",
  "codecov>=2.1.13,<3",
  "isort[colors]>=5.13.2,<6",
  "mypy>=1.10.0,<2",
  "mypy-extensions>=1.0.0,<2",
  "pre-commit>=3.7.1,<4",
  "pytest>=8.2.1,<9",
  "pytest-cov>=5.0.0,<6",
  "pytest-mock>=3.14.0,<4",
  "pyupgrade>=3.15.2,<4",
]
docs = [
  "myst-parser>=3.0.1,<4",
  "numpydoc>=1.7.0,<2",
  "sphinx>=7.3.7,<8",
  "sphinx-autodoc-typehints>=2.1.0,<3",
  "sphinx-click>=6.0.0,<7",
  "sphinx-copybutton>=0.5.2,<1",
  "sphinxcontrib-mermaid>=1.0.0,<2",
  "sphinx-rtd-theme>=2.0.0,<3",
]

[tool.hatch.build.targets.wheel]
packages = ["src/bsr", "src/elastica_blender", "src/virtual_field"]

[tool.black]
# https://github.com/psf/black
target-version = ["py310"]
line-length = 80
color = true

exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | env
    | venv
)/
'''

[tool.isort]
# https://github.com/timothycrosley/isort/
line_length = 80

known_typing = ["typing", "types", "typing_extensions", "mypy", "mypy_extensions"]
sections = ["FUTURE", "TYPING", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
include_trailing_comma = true
profile = "black"
multi_line_output = 3
indent = 4
color_output = true

[tool.mypy]
# https://mypy.readthedocs.io/en/latest/config_file.html#using-a-pyproject-toml-file
python_version = "3.10"
pretty = true
show_traceback = true
color_output = true
strict = true

allow_redefinition = false
check_untyped_defs = false
disallow_any_generics = false
disallow_incomplete_defs = true
disallow_subclassing_any = false # Needed for allowing Elastica mixins
disallow_untyped_calls = true
ignore_missing_imports = true
implicit_reexport = true
no_implicit_optional = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
strict_equality = true
strict_optional = true
warn_no_return = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = false  # TODO: open to discussion
warn_unused_configs = true
warn_unused_ignores = true

[tool.pytest.ini_options]
# https://docs.pytest.org/en/6.2.x/customize.html#pyproject-toml
# Directories that are not visited by pytest collector:
norecursedirs =["hooks", "*.egg", ".eggs", "dist", "build", "docs", ".tox", ".git", "__pycache__"]
doctest_optionflags = ["NUMBER", "NORMALIZE_WHITESPACE", "IGNORE_EXCEPTION_DETAIL", "ELLIPSIS"]

# Extra options:
addopts = [
  "--strict-markers",
  "--tb=short",
#  "--doctest-modules", TODO: Try to include pytest.doctest
#  "--doctest-continue-on-failure",
]

[tool.coverage.report]
# Regexes for lines to exclude from consideration
exclude_lines = [
    # Enable pragma
    "pragma: no cover",
    "TODO",

    # Don't complain if non-runnable code isn't run:
    "if 0:",
    "if __name__ == __main__:",
    "def __repr__",
    '''def plot_'.*\b':''',
    "if self.debug:",
    "if settings.DEBUG",
    "if TYPE_CHECKING:",
    "raise AssertionError",
    "raise NotImplementedError",
    '''class '.*\bProtocol\)':''',
    ''''@(abc\.)?'abstractmethod''',
    '\.\.\.'
]
fail_under = 40
show_missing = true


[tool.coverage.run]
branch = true
omit = [
    "*/.local/*",
	"setup.py",
    "*/protocol.py",
    "macros.py",
]
